requirements.txt
Flask
Flask-Session
sqlite3
pillow
datetime